This is the addition of QA
Inside the QA folder is the QA langchain GPT-4 implementation. There is a txt file where news/information goes in and you can output with customized prompts. There are detailed instructions in qa.py on understanding how to use QA and how to implement it into other models. I plan to add code to make it easier to add a news article into the txt file instead of copying and pasting the whole thing.